Chart gets data through props
The parent componenet of chart is where data is stored now and data is passed to the chart through props.
